readme
This repository contains an Exploratory Data Analysis (EDA) of the iconic TV show Friends. The goal of this analysis is to uncover interesting insights and trends about the series by examining various aspects such as episode ratings, character appearances, dialogue, and more. This analysis aims to provide a deeper understanding of the show’s structure, popularity, and character dynamics.